+++ lib/src/api/model/notification.dart
@@ -1,6 +1,7 @@
 import 'package:InstiApp/src/api/model/event.dart';
 import 'package:InstiApp/src/api/model/post.dart';
 import 'package:InstiApp/src/api/model/serializers.dart';
+import 'package:InstiApp/src/api/model/venter.dart';
 import 'package:jaguar_serializer/jaguar_serializer.dart';
 
 part 'notification.jser.dart';
@@ -8,6 +9,7 @@ part 'notification.jser.dart';
 const String TYPE_EVENT = "event";
 const String TYPE_NEWSENTRY = "newsentry";
 const String TYPE_BLOG = "blogentry";
+const String TYPE_COMPLAINT_COMMENT = "complaintcomment";
 
 class Notification {
   @Alias("id")
@@ -32,6 +34,9 @@ class Notification {
       return getNews().title;
     } else if (isBlogPost) {
       return getBlogPost().title;
+    } else if (isComplaintComment) {
+      var comment = getComment();
+      return "\"${comment.text}\" by ${comment.commentedBy.userName}";
     }
     return "Notification";
   }
@@ -46,6 +51,8 @@ class Notification {
       return ev.eventImageURL ?? ev.eventBodies[0].bodyImageURL;
     } else if (isNews) {
       return getNews().body.bodyImageURL;
+    } else if (isComplaintComment) {
+      return getComment().commentedBy.userProfilePictureUrl;
     }
     return null;
   }
@@ -55,6 +62,9 @@ class Notification {
   bool get isNews => notificationActorType.contains(TYPE_NEWSENTRY);
 
   bool get isBlogPost => notificationActorType.contains(TYPE_BLOG);
+
+  bool get isComplaintComment =>
+      notificationActorType.contains(TYPE_COMPLAINT_COMMENT);
 
   Event getEvent() {
     return standardSerializers.oneFrom<Event>(notificationActor);
@@ -68,6 +78,10 @@ class Notification {
     return standardSerializers.oneFrom<Post>(notificationActor);
   }
 
+  Comment getComment() {
+    return standardSerializers.oneFrom<Comment>(notificationActor);
+  }
+
   String getID() {
     if (isEvent) {
       return getEvent().eventID;
@@ -75,6 +89,8 @@ class Notification {
       return getNews().postID;
     } else if (isBlogPost) {
       return getBlogPost().postID;
+    } else if (isComplaintComment) {
+      return getComment().id;
     }
     return "";
   }