+++ lib/src/blocs/ia_bloc.dart
@@ -275,8 +275,8 @@ class InstiAppBloc {
         : client.getUser(getSessionIdHeader(), uuid);
   }
 
-  Future<Complaint> getComplaint(String uuid) async {
-    return complaintsBloc.getComplaint(uuid);
+  Future<Complaint> getComplaint(String uuid, {bool reload = false}) async {
+    return complaintsBloc.getComplaint(uuid, reload: reload);
   }
 
   // Section