+++ Changelog.md
@@ -4,9 +4,11 @@
 
 * Add back buttons on every page for iOS
 * Add launcher/app icons for iOS
-* Complaint subscription added
+* Complaint subscription added and comments notifications work
 * Move dropdown button to FAB in Mess Menu
 * Fix multiple roles (Hero animation) bug
+* Add InstiApp launcher icon on Android
+* Can edit/delete comments on a Complaint
 
 ## v1.5-beta